Adds byte to the type of cases in the JsonFormatter, which goes through all the fields that were modified in the query and parses them into an value changestream understands.
Changestream relies on https://github.com/shyiko/mysql-binlog-connector-java/tree/dd710a5466381faa57442977b24fceff56a0820e as the main serializer of events that happen in the binlog.

Due to optimizations, they serialize text fields as byte arrays, but changestream was not logging these into the change message itself. This defeats the purpose for a lot of fields in subscribed events and eventually audit logs